First look at 0.9 changes. WIP. Not a working branch. READ-ONLY.
(Pro tip: Don't try to read the diff, it's big.)

This branch is in a disfunctional state. I may update it occasionally as
0.9 gets closer to being functional again. This branch WILL BE DELETED
before 0.9 is ready for trial use. A new branch will be created with a
working copy that will eventually be fit to merge into master, at which
point the beta releases will begin.

None of the middlewares have been ported over yet. That will come close
to last. Main focus right now is laying all the groundwork with a design
that will stand the test of time. That means Caddy must be as extensible
as possible as easily as possible. I've surgically separated most of the
HTTPS and TLS code so that TLS can be used for more than just HTTP.
I'm also completley revamping virtual hosting.

Lots and lots of commented code here, lots and lots of TODOs. Lots of
tests are missing as well. Don't panic. :) This will all be worked out
before 0.9 is released. Code will be cleaned up substantially.

Your feedback is welcome. Thanks!!