Updated look at Caddy 0.9 work. WIP! Not a working branch.
Don't read the diff; it's huge.

You can clone this branch to $GOPATH/src/github.com/mholt/caddy2 and
view the godoc using the `godoc -http=:6060` command, then open your
browser. Otherwise, feel free to browse the code and provide feedback
on the forum or in issues but remember that this is very much a WIP,
lots of stuff is still broken, but lots of stuff is starting to work
too, which is exciting!

Lately I've been focusing a lot on graceful reloads, auto HTTPS + ACME
challenges, and moving the main function to a separate function so you
can have go-gettable custom builds (thanks to captncraig's suggestion).

Further code cleanup, more tests, and many, many fixes and unimplemented
functionality still to come. Looking promising so far!