Initial implementation of 'if' statements for Caddyfile
This a rough, but working, implementation of the 'if' directive (kind of
a faux-directive hacked together in the HTTP server core). It allows any
directive which supports it to conditionally execute based on params of
the request. This is a WIP and needs lots of attention still.

The plumbing is mostly complete, but only the 'status' directive has
been upgraded to support this RequestMatcher. All standard directives
that add an HTTP middleware should be upgraded to support it. Directives
that implemented their own 'if' matching (like 'rewrite', etc) will need
to use this instead.

Other than that, tests need to be added, plugin boilerplate code on the
wiki needs to be updated, and the final interface / usage details will
need to be ironed out, especially in the httpserver package.

It may be useful to implement more condition operators too, so that
we can condition on time of day, etc. We also need to add support for
regular expression captures so they can be used by the middleware
within it...

/cc @abiosoft